+++ src/showcase/examples/index.ts
@@ -19,7 +19,7 @@ export * from './LoadingIndicatorPage';
 export * from './ModalPage';
 export * from './SnackbarPage';
 export * from './TablePage';
-export * from './TabsetPage';
+export * from './TabsPage';
 export * from './TooltipPage';
 export * from './TreeViewPage';
 export * from './TypographyPage';